feat: can change the item position on inventory
for now can only change to an empty slot. Also we have a couple of bugs: when pressing spacebar it automatically grabs an item. When having an item in hand, you can pick an item and the one you had on the hand gets duplicated. Ill try to fix it soon.